mcq_contest/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cors==1.0.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
filelock==3.18.0
future==1.0.0
gevent==25.4.1
google-auth==2.39.0
greenlet==3.2.1
idna==3.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.12.0
PySocks==1.7.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
setuptools==79.0.0
sqlparse==0.5.3
tldextract==5.3.0
tzdata==2025.2
urllib3==2.4.0
zope.event==5.0
zope.interface==7.2
